Moved formatting functions into their own trait
